+++ mne_denoise/asr/core.py
@@ -13,10 +13,17 @@ import warnings
 from typing import Any
 
 import numpy as np
-from scipy import signal, special
+from scipy import special
 from sklearn.base import BaseEstimator, TransformerMixin
 
 from ..utils import extract_data_from_mne, reconstruct_mne_object
+from ._filters import (
+    _append_clean_rawdata_tail,
+    _apply_statistics_filter,
+    _apply_statistics_filter_streaming,
+    _design_statistics_filter,
+    _prepend_clean_rawdata_carry,
+)
 
 # SPD / Riemannian primitives live in mne_denoise.asr._spd.
 from ._spd import (
@@ -28,6 +35,14 @@ from ._spd import (
     _sqrtm_spd,
 )
 from ._types import ASRState
+from ._validation import (
+    _check_enough_samples,
+    _resolve_max_dims,
+    _resolve_max_dims_clean_rawdata,
+    _round_half_up,
+    _validate_array_2d,
+    _validate_common_params,
+)
 
 try:
     import mne
@@ -1898,97 +1913,6 @@ class ASR(BaseEstimator, TransformerMixin):
             raise RuntimeError("ASR is not fitted. Call fit() first.")
 
 
-def _validate_common_params(
-    *,
-    sfreq: float,
-    cutoff: float,
-    window_length: float,
-    window_overlap: float,
-    max_dropout_fraction: float,
-    min_clean_fraction: float,
-    regularization: float,
-) -> None:
-    if sfreq <= 0:
-        raise ValueError("sfreq must be positive")
-    if cutoff <= 0:
-        raise ValueError("cutoff must be positive")
-    if window_length <= 0:
-        raise ValueError("window_length must be positive")
-    if not (0 <= window_overlap < 1):
-        raise ValueError("window_overlap must be in [0, 1)")
-    if not (0 <= max_dropout_fraction < 1):
-        raise ValueError("max_dropout_fraction must be in [0, 1)")
-    if not (0 < min_clean_fraction <= 1):
-        raise ValueError("min_clean_fraction must be in (0, 1]")
-    if max_dropout_fraction + min_clean_fraction >= 1:
-        raise ValueError(
-            "max_dropout_fraction + min_clean_fraction must be less than 1"
-        )
-    if regularization <= 0:
-        raise ValueError("regularization must be positive")
-
-
-def _validate_array_2d(X: np.ndarray) -> np.ndarray:
-    X = np.asarray(X, dtype=np.float64)
-    if X.ndim != 2:
-        raise ValueError(f"ASR expects a 2D array (n_channels, n_times), got {X.shape}")
-    if X.shape[0] < 2:
-        raise ValueError("ASR requires at least two channels")
-    finite_fraction = np.isfinite(X).mean(axis=1)
-    if np.any(finite_fraction < 0.99):
-        bad = np.where(finite_fraction < 0.99)[0].tolist()
-        raise ValueError(f"Channels contain too many non-finite samples: {bad}")
-    X = np.nan_to_num(X, copy=True)
-    variances = np.var(X, axis=1)
-    max_var = float(np.max(variances))
-    # Relative floor so legitimately small-amplitude data (e.g. MEG in Tesla,
-    # variance ~1e-26) is not rejected, while genuinely flat/dead channels
-    # (variance ~0 relative to the rest) still are.
-    if max_var <= 0.0:
-        raise ValueError("All channels have zero or near-zero variance")
-    bad = np.where(variances <= max_var * 1e-12)[0].tolist()
-    if bad:
-        raise ValueError(f"Channels with zero or near-zero variance: {bad}")
-    return X
-
-
-def _check_enough_samples(n_times: int, sfreq: float, window_length: float) -> None:
-    n_win = int(round(window_length * sfreq))
-    if n_win < 2:
-        raise ValueError("window_length is too short for the sampling frequency")
-    if n_times < n_win:
-        raise ValueError(
-            f"Window length ({n_win} samples) exceeds data length ({n_times} samples)"
-        )
-
-
-def _design_statistics_filter(
-    sfreq: float,
-    filter_kind: str,
-) -> tuple[np.ndarray, np.ndarray]:
-    if filter_kind == "none":
-        return np.array([1.0]), np.array([1.0])
-    if filter_kind not in ("asr", "highpass"):
-        raise ValueError("filter_kind must be 'none', 'asr', or 'highpass'")
-    cutoff = min(0.5, sfreq * 0.1)
-    if cutoff >= sfreq / 2:
-        return np.array([1.0]), np.array([1.0])
-    return signal.butter(2, cutoff, btype="highpass", fs=sfreq)
-
-
-def _apply_statistics_filter(
-    X: np.ndarray,
-    b: np.ndarray,
-    a: np.ndarray,
-) -> np.ndarray:
-    if b.size == 1 and a.size == 1:
-        return X.copy()
-    padlen = 3 * max(len(a), len(b))
-    if X.shape[1] <= padlen:
-        return signal.lfilter(b, a, X, axis=1)
-    return signal.filtfilt(b, a, X, axis=1)
-
-
 def _window_starts(n_times: int, win_len: int, overlap: float) -> np.ndarray:
     if win_len < 2:
         raise ValueError("window length must be at least 2 samples")
@@ -2037,11 +1961,6 @@ def _window_rms(X: np.ndarray, starts: np.ndarray, win_len: int) -> np.ndarray:
     return np.sqrt(np.sum(squared, axis=2) / win_len)
 
 
-def _round_half_up(value: float) -> int:
-    """Round non-negative values like MATLAB ``round``."""
-    return int(np.floor(float(value) + 0.5))
-
-
 def _resolve_max_bad_channels_count(
     max_bad_channels: float | int,
     n_channels: int,
@@ -2054,49 +1973,6 @@ def _resolve_max_bad_channels_count(
     if resolved < 0:
         raise ValueError("max_bad_channels must be non-negative")
     return min(resolved, n_channels)
-
-
-def _resolve_max_dims_clean_rawdata(max_dims: float | int, n_channels: int) -> int:
-    """Resolve ASR ``maxdims`` using clean_rawdata's convention."""
-    if isinstance(max_dims, float) and max_dims < 1:
-        if max_dims < 0:
-            raise ValueError("max_dims must be non-negative")
-        return _round_half_up(n_channels * max_dims)
-    max_dims_int = int(max_dims)
-    if max_dims_int < 0:
-        raise ValueError("max_dims must be non-negative")
-    return min(max_dims_int, n_channels)
-
-
-def _append_clean_rawdata_tail(X: np.ndarray, n_tail: int) -> np.ndarray:
-    """Append the reflected lookahead tail used by ``clean_asr.m``."""
-    if n_tail == 0:
-        return X.copy()
-    if X.shape[1] <= n_tail:
-        raise ValueError("Data must contain more samples than the lookahead tail")
-    tail = 2.0 * X[:, [-1]] - X[:, -2 : -n_tail - 2 : -1]
-    return np.concatenate([X, tail], axis=1)
-
-
-def _prepend_clean_rawdata_carry(X: np.ndarray, n_carry: int) -> np.ndarray:
-    """Prepend the reflected initial carry used by ``asr_process.m``."""
-    if n_carry == 0:
-        return X.copy()
-    if X.shape[1] <= n_carry:
-        raise ValueError("Data must contain more samples than the lookahead carry")
-    carry = 2.0 * X[:, [0]] - X[:, n_carry:0:-1]
-    return np.concatenate([carry, X], axis=1)
-
-
-def _apply_statistics_filter_streaming(
-    X: np.ndarray,
-    b: np.ndarray,
-    a: np.ndarray,
-) -> np.ndarray:
-    """Apply the statistics filter in the causal direction used by ASR."""
-    if b.size == 1 and a.size == 1:
-        return X.copy()
-    return signal.lfilter(b, a, X, axis=1)
 
 
 def _process_memory_info(
@@ -2800,17 +2676,6 @@ def _robust_location_scale(values: np.ndarray) -> tuple[float, float]:
     return mu, sigma
 
 
-def _resolve_max_dims(max_dims: float | int, n_channels: int) -> int:
-    if isinstance(max_dims, float):
-        if not (0 <= max_dims <= 1):
-            raise ValueError("float max_dims must be in [0, 1]")
-        return int(np.floor(max_dims * n_channels))
-    max_dims = int(max_dims)
-    if not (0 <= max_dims <= n_channels):
-        raise ValueError("integer max_dims must be in [0, n_channels]")
-    return max_dims
-
-
 def _good_raw_sample_mask(raw: Any, prefixes: tuple[str, ...]) -> np.ndarray:
     n_times = raw.n_times
     mask = np.ones(n_times, dtype=bool)
@@ -2856,3 +2721,18 @@ def _mask_to_sample_spans(mask: np.ndarray) -> list[tuple[int, int]]:
     starts = np.flatnonzero(edges == 1)
     stops = np.flatnonzero(edges == -1)
     return list(zip(starts.tolist(), stops.tolist()))
+
+
+__all__ = [
+    "_validate_common_params",
+    "_validate_array_2d",
+    "_check_enough_samples",
+    "_round_half_up",
+    "_resolve_max_dims_clean_rawdata",
+    "_resolve_max_dims",
+    "_design_statistics_filter",
+    "_apply_statistics_filter",
+    "_append_clean_rawdata_tail",
+    "_prepend_clean_rawdata_carry",
+    "_apply_statistics_filter_streaming",
+]
